2004-03-21 00:30 UTC+0100 Przemyslaw Czerpak (druzus/at/priv.onet.pl)   * xharbour/source/rdd/dbfcdx/dbfcdx1.c     ! fixed smal typo but with very bad side effect - when only bottom       scope was set it cause GPF when skipping     * reorganized TagBOF/TagEOF flags for better performance and fixing       some possible bad BOF/EOF setting     * cdxSkipRaw() is now real SKIPRAW() without any filtering other       then the one set on CDX level.
  * xharbour/source/rdd/workarea.c
    * fixed my last modification, now skipFilter() do only minimum
      necessary operations

  * xharbour/source/rdd/dbfntx/dbfntx1.c
    + flushing buffers added (patch from "Oscar Hernandez Suarez"
      <oscar@oscarweb.com>, thanks)

  - xharbour/tests/bldtest/endian.c
  + xharbour/tests/bldtest/bldtest.c
    * endian.c renamed to bldtest.c and a little bit extend

  * xharbour/include/hbexprb.c
  * xharbour/source/common/hbver.c
  * xharbour/source/pp/ppcore.c
  * xharbour/source/rtl/at.c
  * xharbour/source/rtl/filesys.c
    * warning fixes

+++ xharbour/source/rtl/at.c
@@ -1,5 +1,5 @@
 /*
- * $Id: at.c,v 1.3 2002/05/08 20:07:07 ronpinkas Exp $
+ * $Id: at.c,v 1.4 2003/05/17 17:30:28 jonnymind Exp $
  */
 
 /*
@@ -84,7 +84,7 @@
        {
           ULONG ulLength = pText->item.asString.length;
           ULONG ulStart = ISNUM( 3 ) ? hb_parnl( 3 ) : 1;
-          ULONG ulEnd = ISNUM( 4 ) ? hb_parnl( 4 ) : ulLength;
+          ULONG ulEnd = ISNUM( 4 ) ? (ULONG) hb_parnl( 4 ) : ulLength;
           ULONG ulPos;
 
           // SANITIZATION
